+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 1.2.0
+- Install action added
+- Uninstall action added
+- Upgrade action added
+- getCatalogSites action added
+
 ## 1.1.0
 - Retract action added
 - Remove action added